# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpportuneX API
  description: |
    OpportuneX is an AI-powered platform that helps students from Tier 2 and Tier 3 cities in India discover and access hackathons, internships, and workshops through natural language interactions.
    
    ## Features
    - **Smart Search**: Natural language search for opportunities with advanced filtering
    - **Voice Search**: Multi-language voice commands (English and Hindi)
    - **AI Instructor**: Personalized preparation roadmaps and guidance
    - **Smart Notifications**: Personalized opportunity alerts via email, SMS, and in-app
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General API endpoints: 100 requests per 15 minutes
    - Search endpoints: 50 requests per 15 minutes
    - Authentication endpoints: 10 requests per 15 minutes
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Invalid request parameters",
        "details": ["Field 'query' is required"]
      }
    }
    ```
  version: 1.0.0
  contact:
    name: OpportuneX API Support
    email: api-support@opportunex.com
    url: https://opportunex.com/support
  license:
    name: MIT
...
servers:
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "student@example.com"
              password: "SecurePassword123!"
              name: "John Doe"
              location:
                city: "Indore"
                state: "Madhya Pradesh"
                tier: 2
              academic:
                institution: "IIT Indore"
                degree: "B.Tech Computer Science"
                year: 3
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "student@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Get a new JWT token using refresh token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Search Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - location
        - academic
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password (minimum 8 characters)
        name:
          type: string
          minLength: 2
          description: User's full name
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
          description: User's phone number (optional)
        location:
          $ref: '#/components/schemas/Location'
        academic:
          $ref: '#/components/schemas/Academic'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              description: JWT access token
            refreshToken:
              type: string
              description: Refresh token
            expiresIn:
              type: integer
              description: Token expiration time in seconds

    # Search Schemas
...
    SearchFilters:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
          description: Required technical skills
        organizerType:
          type: string
          enum: ["corporate", "startup", "government", "academic"]
          description: Type of organizing entity
        mode:
          type: string
          enum: ["online", "offline", "hybrid"]
          description: Event mode
        location:
          type: string
          description: Geographic location
        type:
          type: string
          enum: ["hackathon", "internship", "workshop"]
          description: Opportunity type
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        academic:
          $ref: '#/components/schemas/Academic'
        skills:
          $ref: '#/components/schemas/Skills'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        searchHistory:
          type: array
          items:
            $ref: '#/components/schemas/SearchQuery'
        favoriteOpportunities:
          type: array
          items:
            type: string
          description: Array of favorite opportunity IDs
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Location:
      type: object
      required:
        - city
        - state
        - tier
      properties:
        city:
          type: string
          description: City name
        state:
          type: string
          description: State name
        tier:
          type: integer
          enum: [2, 3]
          description: City tier (2 or 3)
    Academic:
      type: object
      required:
        - institution
        - degree
        - year
      properties:
        institution:
          type: string
          description: Educational institution name
        degree:
          type: string
          description: Degree program
        year:
          type: integer
          minimum: 1
          maximum: 6
          description: Current year of study
        cgpa:
          type: number
          minimum: 0
          maximum: 10
          description: Current CGPA (optional)
    Skills:
      type: object
      properties:
        technical:
          type: array
          items:
            type: string
          description: Technical skills
        domains:
          type: array
          items:
            type: string
          description: Domain expertise areas
        proficiencyLevel:
          type: string
          enum: ["beginner", "intermediate", "advanced"]
          description: Overall proficiency level
    UserPreferences:
      type: object
      properties:
        opportunityTypes:
          type: array
          items:
            type: string
            enum: ["hackathon", "internship", "workshop"]
          description: Preferred opportunity types
        preferredMode:
          type: string
          enum: ["online", "offline", "hybrid", "any"]
          description: Preferred event mode
        maxDistance:
          type: integer
          minimum: 0
          description: Maximum distance for offline events (km)
        notifications:
          $ref: '#/components/schemas/NotificationPreferences'
    NotificationPreferences:
      type: object
      properties:
        email:
          type: boolean
          description: Enable email notifications
        sms:
          type: boolean
          description: Enable SMS notifications
        push:
          type: boolean
          description: Enable push notifications
        frequency:
          type: string
          enum: ["immediate", "daily", "weekly"]
          description: Notification frequency
        types:
          type: array
          items:
            type: string
            enum: ["opportunity_match", "deadline_reminder", "roadmap_update"]
          description: Types of notifications to receive

    # AI Instructor Schemas
...
    SearchQuery:
      type: object
      properties:
        query:
          type: string
        filters:
          $ref: '#/components/schemas/SearchFilters'
        timestamp:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: string
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request parameters"
              details: ["Field 'query' is required"]
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired token"
...
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "USER_EXISTS"
              message: "User with this email already exists"
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Too many requests. Please try again later."
...